update ui: user management

+++ src/admin/users/UserManager.vue
@@ -1,129 +1,212 @@
 <template>
-  <div style="padding-top: 30px;" class="user-manage">
-    <div style="display: flex;">
-      <h2>User Management</h2>
-      <div style="height:80px;padding-top: 10px;">
-        <Button type="primary" @click="newModal = true">Create new user</Button>
+  <div class="user-manage">
+    <Panel title="User management">
+      <div slot="header">
+        <el-col :span="24">
+          <el-input v-model="keyword" placeholder="Keywords"
+                    prefix-icon="el-icon-search"></el-input>
+        </el-col>
       </div>
-    </div>
-    <Row>
-      <Col span="24">
-        <ul style="font-weight: 700" class="pro-table">
-          <li class="name">Full name</li>
-          <li class="username">Username</li>
-          <li class="mail">Mail</li>
-          <li class="diff">Accept count</li>
-          <li class="diff">AC rate</li>
-          <li class="rate">Role</li>
-          <li class="rate">Modify</li>
-        </ul>
-        <ul
-            v-for="(user, index) in users"
-            :key="user.id"
-            class="pro-table"
-            :class="[index % 2 === 0 ? 'bg' : '']"
-        >
-          <li class="name">{{ user.name }}</li>
-          <li class="username">{{ user.username }}</li>
-          <li class="mail">{{ user.email }}</li>
-          <li class="diff">{{ user.acCount}}</li>
-          <li class="diff">{{ (user.acRate * 100).toFixed(2) }}%</li>
-          <li class="rate">{{ getRole(user) }}</li>
-          <li class="rate">
-            <Button type="primary" class="btn-primary" @click="reviseInfo(user)"
-            >Modify</Button
-            >
-            <!--<Button type="danger" class="btn-primary" @click="deleteUser(user.id)">Delete</Button>-->
-          </li>
-        </ul>
-      </Col>
-      <Modal
-          v-model="newModal"
-          title="Create new user"
-          width="50%"
-          @on-ok="createUser"
-          @on-cancel="newModal = false"
-      >
-        <div style="display: flex;">
-          <div
-              style="display: flex;height: 220px;padding-top:10px;width: 60px;flex-direction: column;justify-content: space-between"
-          >
-            <div>Username</div>
-            <div>Password</div>
-            <div>Email</div>
-            <div>First name</div>
-            <div>Last name</div>
-            <div>Role</div>
-          </div>
-          <div>
-            <Input v-model="newUser.username" />
-            <Input v-model="newUser.password" />
-            <Input v-model="newUser.email" />
-            <Input v-model="newUser.firstname" />
-            <Input v-model="newUser.lastname" />
-            <Select v-model="role" style="width:200px">
-              <Option
-                  v-for="item in roleList"
-                  :value="item.value"
-                  :key="item.value"
-              >{{ item.label }}</Option
-              >
-            </Select>
-          </div>
-        </div>
-      </Modal>
-      <Modal
-          v-model="reviseModal"
-          title="Modify user info"
-          width="50%"
-          @on-ok="reviseUserInfo"
-          @on-cancel="reviseModal = false"
-      >
-        <div style="display: flex;">
-          <div
-              style="display: flex;height: 180px;padding-top:10px;width: 50px;flex-direction: column;justify-content: space-between"
-          >
-            <div>Username</div>
-            <div>Email</div>
-            <div>First name</div>
-            <div>Last name</div>
-            <div>Role</div>
-          </div>
-          <div>
-            <Input v-model="reviseUser.username" />
-            <Input v-model="reviseUser.email" />
-            <Input v-model="reviseUser.firstname" />
-            <Input v-model="reviseUser.lastname" />
-            <Input v-model="reviseUser.school" />
-            <Select v-model="role">
-              <Option
-                  v-for="item in roleList"
-                  :value="item.value"
-                  :key="item.value"
-              >{{ item.label }}</Option
-              >
-            </Select>
-          </div>
-        </div>
-      </Modal>
-      <Col span="24" class="card-margin">
-        <Page
-            :total="total"
-            show-sizer
-            @on-change="pageChange"
-            @on-page-size-change="pageSizeChange"
+
+      <el-table
+          ref="table"
+          v-loading="loadingTable"
+          :data="users"
+          element-loading-text="loading"
+          style="width: 100%">
+        <el-table-column label="ID" prop="id"></el-table-column>
+
+        <el-table-column label="Username" prop="username"></el-table-column>
+
+        <el-table-column label="Create Time" prop="create_time">
+          <template slot-scope="scope">
+            {{ scope.row.createDate }}
+          </template>
+        </el-table-column>
+
+        <el-table-column label="Full Name" prop="name"></el-table-column>
+        <el-table-column label="AC" prop="acCount"></el-table-column>
+        <el-table-column label="Email" prop="email"></el-table-column>
+
+        <el-table-column label="User Type">
+          <template slot-scope="scope">
+            {{ getRole(scope.row.authorities[0].name) }}
+          </template>
+        </el-table-column>
+
+        <el-table-column fixed="right" label="Option" width="200">
+          <template slot-scope="{row}">
+            <Button icon="md-create" name="Edit" @click.native="openUserDialog(row.id)"></Button>
+            <Button icon="md-trash" name="Delete" @click.native="deleteUser(row.id)"></Button>
+          </template>
+        </el-table-column>
+      </el-table>
+      <div class="panel-options">
+        <el-button type="primary" size="small"
+                   @click="goCreateAdminUser" icon="el-icon-plus">Create Admin
+        </el-button>
+        <Page class="page"
+              :total="total"
+              show-sizer
+              @on-change="pageChange"
+              @on-page-size-change="pageSizeChange"
         />
-      </Col>
-    </Row>
+      </div>
+    </Panel>
+    <!--      <Modal-->
+    <!--          v-model="newModal"-->
+    <!--          title="Create new user"-->
+    <!--          width="50%"-->
+    <!--          @on-ok="createUser"-->
+    <!--          @on-cancel="newModal = false"-->
+    <!--      >-->
+    <!--        <div style="display: flex;">-->
+    <!--          <div-->
+    <!--              style="display: flex;height: 220px;padding-top:10px;width: 60px;flex-direction: column;justify-content: space-between"-->
+    <!--          >-->
+    <!--            <div>Username</div>-->
+    <!--            <div>Password</div>-->
+    <!--            <div>Email</div>-->
+    <!--            <div>First name</div>-->
+    <!--            <div>Last name</div>-->
+    <!--            <div>Role</div>-->
+    <!--          </div>-->
+    <!--          <div>-->
+    <!--            <Input v-model="newUser.username" />-->
+    <!--            <Input v-model="newUser.password" />-->
+    <!--            <Input v-model="newUser.email" />-->
+    <!--            <Input v-model="newUser.firstname" />-->
+    <!--            <Input v-model="newUser.lastname" />-->
+    <!--            <Select v-model="role" style="width:200px">-->
+    <!--              <Option-->
+    <!--                  v-for="item in roleList"-->
+    <!--                  :value="item.value"-->
+    <!--                  :key="item.value"-->
+    <!--              >{{ item.label }}</Option-->
+    <!--              >-->
+    <!--            </Select>-->
+    <!--          </div>-->
+    <!--        </div>-->
+    <!--      </Modal>-->
+
+    <el-dialog :close-on-click-modal="false" :visible.sync="showUserDialog" title="User Info">
+      <el-form :model="user" label-position="left" label-width="120px">
+        <el-row :gutter="20">
+          <el-col :span="12">
+            <el-form-item label="Username" required>
+              <el-input v-model="user.username"></el-input>
+            </el-form-item>
+          </el-col>
+          <el-col :span="12">
+            <el-form-item label="Name" required>
+              <el-input v-model="user.name"></el-input>
+            </el-form-item>
+          </el-col>
+          <el-col :span="12">
+            <el-form-item label="Email" required>
+              <el-input v-model="user.email"></el-input>
+            </el-form-item>
+          </el-col>
+          <el-col :span="12">
+            <el-form-item label="User type">
+              <el-select v-model="role">
+                <el-option label="User" value="ROLE_USER"></el-option>
+                <el-option label="Admin" value="ROLE_ADMIN"></el-option>
+                <el-option label="Super Admin" value="ROLE_SUPER_ADMIN"></el-option>
+              </el-select>
+            </el-form-item>
+          </el-col>
+          <el-col :span="12">
+            <el-form-item label="Problem Permission">
+              <el-select v-model="user.problemPermission" :disabled="role !== 'ROLE_ADMIN' ">
+                <el-option label="None" value="NONE"></el-option>
+                <el-option label="Own" value="OWN"></el-option>
+                <el-option label="All" value="ALL"></el-option>
+              </el-select>
+            </el-form-item>
+          </el-col>
+          <el-col :span="8">
+            <el-form-item label="Enable">
+              <el-switch
+                  v-model="user.enabled">
+              </el-switch>
+            </el-form-item>
+          </el-col>
+        </el-row>
+      </el-form>
+      <span slot="footer" class="dialog-footer">
+        <Button @click.native="showUserDialog = false">Cancel</Button>
+        <Button @click.native="updateUserInfo()">Save</Button>
+      </span>
+    </el-dialog>
+    <el-dialog :close-on-click-modal="false" :visible.sync="showCreateAdminDialog" title="Create admin account">
+      <el-form :model="newUser" label-position="left" label-width="120px">
+        <el-row :gutter="20">
+          <el-col :span="12">
+            <el-form-item label="First Name" required>
+              <el-input v-model="newUser.firstname"></el-input>
+            </el-form-item>
+          </el-col>
+          <el-col :span="12">
+            <el-form-item label="Last Name" required>
+              <el-input v-model="newUser.lastname"></el-input>
+            </el-form-item>
+          </el-col>
+          <el-col :span="12">
+            <el-form-item label="Username" required>
+              <el-input v-model="newUser.username"></el-input>
+            </el-form-item>
+          </el-col>
+          <el-col :span="12">
+            <el-form-item label="Email" required>
+              <el-input v-model="newUser.email"></el-input>
+            </el-form-item>
+          </el-col>
+          <el-col :span="12">
+            <el-form-item label="Password" required>
+              <el-input v-model="newUser.password"></el-input>
+            </el-form-item>
+          </el-col>
+          <el-col :span="12">
+            <el-form-item label="User type">
+              <el-select v-model="role">
+                <el-option label="User" value="ROLE_USER"></el-option>
+                <el-option label="Admin" value="ROLE_ADMIN"></el-option>
+                <el-option label="Super Admin" value="ROLE_SUPER_ADMIN"></el-option>
+              </el-select>
+            </el-form-item>
+          </el-col>
+          <el-col :span="12">
+            <el-form-item label="Problem Permission">
+              <el-select v-model="newUser.problemPermission" :disabled="role !== 'ROLE_ADMIN' ">
+                <el-option label="None" value="NONE"></el-option>
+                <el-option label="Own" value="OWN"></el-option>
+                <el-option label="All" value="ALL"></el-option>
+              </el-select>
+            </el-form-item>
+          </el-col>
+        </el-row>
+      </el-form>
+      <span slot="footer" class="dialog-footer">
+        <Button @click.native="showUserDialog = false">Cancel</Button>
+        <Button @click.native="createUser()">Save</Button>
+      </span>
+    </el-dialog>
   </div>
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'vue-property-decorator'
+import {Component, Vue} from 'vue-property-decorator'
 import api from '@/api/api'
 import md5 from 'js-md5'
+import Panel from "@/components/Panel.vue";
 
-@Component
+@Component({
+      components: {
+        Panel
+      }
+    }
+)
 export default class UserManager extends Vue {
   users: any = []
   pageSize: number = 10
@@ -131,8 +214,6 @@ export default class UserManager extends Vue {
   total: any = 0
   reviseUser: any = {}
   newUser: any = {}
-  newModal: boolean = false
-  reviseModal: boolean = false
   role: string = ''
   roleList: Array<any> = [
     {
@@ -149,20 +230,25 @@ export default class UserManager extends Vue {
     },
   ]
 
-  reviseInfo(user: any) {
-    this.reviseUser = JSON.parse(JSON.stringify(user))
-    this.reviseModal = true
-  }
+  keyword: string = ''
+  loadingTable: boolean = false
+  showUserDialog: boolean = false
+  user: any = {}
+  authority: string = ''
+  showCreateAdminDialog: boolean = false
+
 
   createUser() {
     api.user
     .createUser({
       ...this.newUser,
-      password: md5(this.newUser.password),
-      authorities: [{ name: this.role }],
+      password: this.newUser.password,
+      authorities: [{name: this.role}],
     })
     .then(() => {
+      this.showCreateAdminDialog = false
       ;(this as any).$Message.success('Create successful')
+      this.getUsers(this.page, this.pageSize)
       this.newUser = {}
     })
     .catch((err: any) => {
@@ -170,34 +256,33 @@ export default class UserManager extends Vue {
     })
   }
 
-  reviseUserInfo() {
-    const user: any = {
-      ...this.reviseUser,
-      authorities: [{ name: this.role }],
+  updateUserInfo() {
+    const user = {
+      ...this.user,
+      authorities: [{name: this.role}],
     }
     api.user.updateUserInfo(user).then(res => {
-      this.reviseUser = {}
+      this.$Message.success("Update successful")
       this.getUsers(this.page, this.pageSize)
+      this.showUserDialog = false
+    }).catch(err => {
+      this.$Message.error(err.data.message)
     })
   }
 
-  getRole(user: any) {
-    const type = user.authorities[0]
-    if (type) {
-      switch (type.authority) {
-        case 'ROLE_ADMIN':
-          return 'Admin'
-        case 'ROLE_USER':
-          return 'User'
-        case 'ROLE_SUPER_ADMIN':
-          return 'Super Admin'
-        default:
-          return ''
-      }
-    } else {
-      return ''
+  getRole(role: string) {
+    switch (role) {
+      case 'ROLE_ADMIN':
+        return 'Admin'
+      case 'ROLE_USER':
+        return 'User'
+      case 'ROLE_SUPER_ADMIN':
+        return 'Super Admin'
+      default:
+        return ''
     }
   }
+
   pageChange(pages: number) {
     this.page = pages - 1
     this.getUsers(pages - 1, this.pageSize)
@@ -209,15 +294,28 @@ export default class UserManager extends Vue {
   }
 
   deleteUser(id: string) {
-    api
-    .deleteUser({
-      list: [id],
+    this.$confirm('Sure to delete the user? The associated resources created by this user will be deleted as well, like problem, contest, announcement, etc.', 'confirm', {
+      type: 'warning'
+    }).then(() => {
+      api
+      .deleteUser({
+        id,
+      })
+      .then(res => {
+        ;(this as any).$Message.success('Delete successful')
+        this.getUsers(this.page, this.pageSize)
+      })
+      .catch((err: any) => {
+        ;(this as any).$Message.error(err.data.message)
+      })
     })
-    .then(res => {
-      ;(this as any).$Message.success('Modify successful')
-    })
-    .catch((err: any) => {
-      ;(this as any).$Message.error(err.data.message)
+  }
+
+  openUserDialog(id: string) {
+    this.showUserDialog = true
+    api.user.getUserInfo({id}).then(res => {
+      this.user = res.data
+      this.role = this.user.authorities[0].name
     })
   }
 
@@ -230,10 +328,15 @@ export default class UserManager extends Vue {
     .then((res: any) => {
       this.users = res.data.list
       this.total = res.data.total
+      console.log(this.users)
     })
     .catch((err: any) => {
       ;(this as any).$Message.error(err.data.message)
     })
+  }
+  goCreateAdminUser(){
+    this.showCreateAdminDialog = true
+    this.newUser = {}
   }
   mounted() {
     this.getUsers(0, 10)
@@ -250,11 +353,13 @@ export default class UserManager extends Vue {
   margin-top: 20px;
   display: flex;
   justify-content: center;
+
   button {
     margin-right: 2px;
     padding: 3px 7px 3px;
   }
 }
+
 .pro-table {
   list-style: none;
   display: flex;
@@ -263,34 +368,43 @@ export default class UserManager extends Vue {
   line-height: 48px;
   text-align: left;
   border-bottom: rgb(221, 221, 221) solid 1px;
+
   .name {
     width: 10%;
     cursor: pointer;
+
     &:hover {
       color: rgb(17, 85, 204);
     }
   }
+
   .username {
     width: 20%;
     cursor: pointer;
+
     &:hover {
       color: rgb(17, 85, 204);
     }
   }
+
   .mail {
     width: 20%;
   }
+
   .diff {
     width: 10%;
   }
+
   .rate {
     width: 20%;
     text-align: center;
   }
+
   .btn-primary {
     margin-left: 10px;
   }
 }
+
 h2 {
   font-weight: 500;
   font-size: 36px;
